♻️ create a base class for azure storage that will be used for classes that manage fileshares & blobs
