+++ clients/azure/__init__.py
@@ -1,2 +1,2 @@
-from .data import StorageAzureClient  # noqa: F401
+from .data import DataAzureClient  # noqa: F401
 from .vm import VMAzureClient  # noqa: F401
