+++ clients/azure/data.py
@@ -4,10 +4,7 @@ from datetime import datetime, timedelta
 from pathlib import Path
 from typing import Generator, Literal, Optional
 
-from azure.core.credentials import TokenCredential
 from azure.core.exceptions import ResourceNotFoundError
-from azure.identity import DefaultAzureCredential
-from azure.mgmt.storage import StorageManagementClient
 from azure.storage.file.models import FilePermissions
 from azure.storage.file.sharedaccesssignature import FileSharedAccessSignature
 from azure.storage.fileshare import (
@@ -20,6 +17,8 @@ from dotenv import load_dotenv
 from pydantic import BaseModel
 
 from auth import User
+
+from ._storage import BaseStorageAzureClient
 
 load_dotenv()
 
@@ -48,25 +47,11 @@ class ProjectFile(BaseModel):
     path: Optional[str]
 
 
-class StorageAzureClient:
+class DataAzureClient(BaseStorageAzureClient):
     def __init__(self):
-        credentials = DefaultAzureCredential()
-        self.resource_group_name = os.environ["AZURE_RESOURCE_GROUP_NAME"]
-        self.storage_account_name = os.environ["AZURE_STORAGE_ACCOUNT"]
-
-        storage_key = _get_storage_key(
-            credentials,
-            os.environ["AZURE_SUBSCRIPTION_ID"],
-            resource_group_name=self.resource_group_name,
-            storage_account_name=self.storage_account_name,
-        )
-
-        # pylint: disable=line-too-long,consider-using-f-string
-        self._storage_connection_string = "DefaultEndpointsProtocol=https;AccountName={};AccountKey={};EndpointSuffix=core.windows.net".format(
-            self.storage_account_name, storage_key
-        )
+        super().__init__()
         self._file_shared_access_signature = FileSharedAccessSignature(
-            account_name=self.storage_account_name, account_key=storage_key
+            account_name=self.storage_account_name, account_key=self._storage_key
         )
 
     def get_project_documents(
@@ -263,25 +248,6 @@ def validate_project_document_file_path(path: Path, current_user: User):
             "path must start with {projects_path_prefix}/<project_name>/documents/"
         )
     _validate_project_file_path(path, current_user)
-
-
-def _get_storage_key(
-    credential: TokenCredential,
-    subscription_id: str,
-    resource_group_name: str,
-    storage_account_name: str,
-):
-    """Fetches a storage account key to use as a credential"""
-    storage_mgmt_client = StorageManagementClient(credential, subscription_id)
-    key = (
-        storage_mgmt_client.storage_accounts.list_keys(
-            resource_group_name, storage_account_name
-        )
-        .keys[0]
-        .value
-    )
-
-    return key
 
 
 def _get_projects_path():
